refactor: change agenda/presentation file structure
