+++ androidApp/src/main/java/com/example/tasky/android/agenda/di/AgendaDetailsModule.kt
@@ -2,7 +2,7 @@ package com.example.tasky.android.agenda.di
 
 import com.example.tasky.android.agenda.data.ImageCompressor
 import com.example.tasky.android.agenda.domain.IImageCompressor
-import com.example.tasky.android.agenda.presentation.viewmodel.AgendaDetailsViewModel
+import com.example.tasky.android.agenda.presentation.details.screen.AgendaDetailsViewModel
 import com.example.tasky.common.data.manager.NetworkManager
 import com.example.tasky.common.domain.INetworkManager
 import org.koin.core.module.dsl.factoryOf
