+++ androidApp/src/main/java/com/example/tasky/android/agenda/di/AgendaModule.kt
@@ -2,7 +2,7 @@ package com.example.tasky.android.agenda.di
 
 import com.example.tasky.agenda.data.AgendaRepository
 import com.example.tasky.agenda.domain.IAgendaRepository
-import com.example.tasky.android.agenda.presentation.viewmodel.AgendaViewModel
+import com.example.tasky.android.agenda.presentation.agenda.screen.AgendaViewModel
 import com.example.tasky.auth.data.manager.SessionManager
 import com.example.tasky.auth.domain.ISessionManager
 import org.koin.core.module.dsl.viewModelOf
