+++ androidApp/src/main/java/com/example/tasky/android/agenda/presentation/agenda/component/AgendaCard.kt
@@ -1,4 +1,4 @@
-package com.example.tasky.android.agenda.presentation.components.agenda
+package com.example.tasky.android.agenda.presentation.agenda.component
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
