+++ androidApp/src/main/java/com/example/tasky/android/agenda/presentation/agenda/screen/AgendaViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.tasky.android.agenda.presentation.viewmodel
+package com.example.tasky.android.agenda.presentation.agenda.screen
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
@@ -10,9 +10,6 @@ import com.example.tasky.agenda.domain.model.Event
 import com.example.tasky.agenda.domain.model.RemindAtType
 import com.example.tasky.agenda.domain.model.Reminder
 import com.example.tasky.agenda.domain.model.Task
-import com.example.tasky.android.agenda.presentation.screen.AgendaItemUi
-import com.example.tasky.android.agenda.presentation.screen.AgendaScreenEvent
-import com.example.tasky.android.agenda.presentation.screen.AgendaScreenState
 import com.example.tasky.auth.domain.IAuthRepository
 import com.example.tasky.auth.domain.ISessionManager
 import com.example.tasky.auth.domain.util.getAvatarDisplayName
