+++ androidApp/src/main/java/com/example/tasky/android/agenda/presentation/details/component/DetailsAttendeeSection.kt
@@ -1,4 +1,4 @@
-package com.example.tasky.android.agenda.presentation.components.details
+package com.example.tasky.android.agenda.presentation.details.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
