+++ androidApp/src/main/java/com/example/tasky/android/agenda/presentation/details/component/DetailsEditText.kt
@@ -1,4 +1,4 @@
-package com.example.tasky.android.agenda.presentation.components.details
+package com.example.tasky.android.agenda.presentation.details.component
 
 import androidx.annotation.StringRes
 import androidx.compose.foundation.background
