+++ androidApp/src/main/java/com/example/tasky/android/agenda/presentation/details/screen/AgendaDetailsScreen.kt
@@ -1,4 +1,4 @@
-package com.example.tasky.android.agenda.presentation.screen
+package com.example.tasky.android.agenda.presentation.details.screen
 
 import android.net.Uri
 import androidx.compose.foundation.background
@@ -38,22 +38,20 @@ import com.example.tasky.agenda.domain.model.Reminder
 import com.example.tasky.agenda.domain.model.Task
 import com.example.tasky.android.AGENDA_DETAILS_DEEPLINK
 import com.example.tasky.android.R
-import com.example.tasky.android.agenda.presentation.components.details.DetailsAttendeeSection
-import com.example.tasky.android.agenda.presentation.components.details.DetailsAttendeeSectionTabOption
-import com.example.tasky.android.agenda.presentation.components.details.DetailsDateTimeSection
-import com.example.tasky.android.agenda.presentation.components.details.DetailsDeleteSection
-import com.example.tasky.android.agenda.presentation.components.details.DetailsDescSection
-import com.example.tasky.android.agenda.presentation.components.details.DetailsEditText
-import com.example.tasky.android.agenda.presentation.components.details.DetailsEditTextType
-import com.example.tasky.android.agenda.presentation.components.details.DetailsHeaderSection
-import com.example.tasky.android.agenda.presentation.components.details.DetailsLargePhoto
-import com.example.tasky.android.agenda.presentation.components.details.DetailsPhoto
-import com.example.tasky.android.agenda.presentation.components.details.DetailsPhotoSection
-import com.example.tasky.android.agenda.presentation.components.details.DetailsRemindAtSection
-import com.example.tasky.android.agenda.presentation.components.details.DetailsTitleSection
-import com.example.tasky.android.agenda.presentation.components.details.DetailsTopBar
-import com.example.tasky.android.agenda.presentation.viewmodel.AgendaDetailsOneTimeEvent
-import com.example.tasky.android.agenda.presentation.viewmodel.AgendaDetailsViewModel
+import com.example.tasky.android.agenda.presentation.details.component.DetailsAttendeeSection
+import com.example.tasky.android.agenda.presentation.details.component.DetailsAttendeeSectionTabOption
+import com.example.tasky.android.agenda.presentation.details.component.DetailsDateTimeSection
+import com.example.tasky.android.agenda.presentation.details.component.DetailsDeleteSection
+import com.example.tasky.android.agenda.presentation.details.component.DetailsDescSection
+import com.example.tasky.android.agenda.presentation.details.component.DetailsEditText
+import com.example.tasky.android.agenda.presentation.details.component.DetailsEditTextType
+import com.example.tasky.android.agenda.presentation.details.component.DetailsHeaderSection
+import com.example.tasky.android.agenda.presentation.details.component.DetailsLargePhoto
+import com.example.tasky.android.agenda.presentation.details.component.DetailsPhoto
+import com.example.tasky.android.agenda.presentation.details.component.DetailsPhotoSection
+import com.example.tasky.android.agenda.presentation.details.component.DetailsRemindAtSection
+import com.example.tasky.android.agenda.presentation.details.component.DetailsTitleSection
+import com.example.tasky.android.agenda.presentation.details.component.DetailsTopBar
 import com.example.tasky.android.common.presentation.utils.ObserveAsEvents
 import com.example.tasky.android.common.presentation.utils.showToast
 import com.example.tasky.android.theme.Black
